pyproject.toml
[project]
name = "security_requirements_system"
version = "0.1.0"
description = "security-requirements-system using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.157.0,<1.0.0",
    "weaviate-client>=4.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
]

[project.scripts]
kickoff = "security_requirements_system.main:kickoff"
run_crew = "security_requirements_system.main:kickoff"
plot = "security_requirements_system.main:plot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
